refactor(todo): split service.go below 1000 lines
Move helpers into sibling files in package todo.
No behavior change.

Related: #652
