+++ internal/todo/service.go
@@ -5,7 +5,6 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/google/uuid"
@@ -13,7 +12,6 @@ import (
 	"github.com/douglasdemoura/chroncal/internal/calendaraccess"
 	"github.com/douglasdemoura/chroncal/internal/duration"
 	"github.com/douglasdemoura/chroncal/internal/hydrate"
-	"github.com/douglasdemoura/chroncal/internal/model"
 	"github.com/douglasdemoura/chroncal/internal/storage"
 	"github.com/douglasdemoura/chroncal/internal/timeutil"
 )
@@ -620,172 +618,6 @@ func (s *Service) UpsertByUID(ctx context.Context, p UpsertParams) (Todo, error)
 	return t, nil
 }
 
-// ErrHasOverrides is returned when a delete targets a recurring master
-// todo that has override instances. Use DeleteSeries instead.
-var ErrHasOverrides = fmt.Errorf("todo has overrides: use DeleteSeries to delete the entire series")
-
-// Delete soft-deletes a todo by ID. For a standalone todo it flips
-// deleted_at. For an override it adds EXDATE to the master and
-// soft-deletes the override in the same transaction so undo can reverse
-// both sides. A recurring master with live overrides is rejected.
-// Callers must use DeleteSeries.
-func (s *Service) Delete(ctx context.Context, id int64) error {
-	td, err := s.Get(ctx, id)
-	if err != nil {
-		return err
-	}
-
-	if err := calendaraccess.EnsureWritable(ctx, s.q, td.CalendarID, todoComponent); err != nil {
-		return err
-	}
-	// RDATE-only masters (no RRULE) are recurring too, so guard on either rule
-	// or RDATEs; otherwise their overrides would be orphaned (#415).
-	if (td.RecurrenceRule != "" || td.RDates != "") && td.RecurrenceID == "" {
-		overrides, err := s.q.ListTodoOverridesByUID(ctx, td.UID)
-		if err != nil {
-			return fmt.Errorf("check overrides: %w", err)
-		}
-		if len(overrides) > 0 {
-			return ErrHasOverrides
-		}
-	}
-
-	if td.RecurrenceID == "" {
-		// Tombstone and soft-delete commit together. A failed tombstone write
-		// cannot leave a soft-deleted row whose next sync DELETEs a still-live
-		// server resource (issue #107).
-		tx, err := s.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("begin tx: %w", err)
-		}
-		defer tx.Rollback()
-		qtx := s.q.WithTx(tx)
-		if _, err := storage.CreateTombstoneIfSynced(ctx, tx, td.CalendarID, td.UID); err != nil {
-			return fmt.Errorf("create tombstone: %w", err)
-		}
-		if err := qtx.SoftDeleteTodo(ctx, id); err != nil {
-			return fmt.Errorf("soft-delete todo: %w", err)
-		}
-		if err := storage.MarkResourceDirty(ctx, tx, td.CalendarID, td.UID, "todo"); err != nil {
-			return fmt.Errorf("mark resource dirty: %w", err)
-		}
-		return tx.Commit()
-	}
-
-	if td.RecurrenceID != "" {
-		tx, err := s.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("begin tx: %w", err)
-		}
-		defer tx.Rollback()
-		qtx := s.q.WithTx(tx)
-
-		master, err := qtx.GetTodoByUID(ctx, td.UID)
-		// A genuine lookup error (for example SQLITE_BUSY) must not collapse
-		// into the "no master" path. That path would soft-delete the override
-		// and skip its EXDATE/provenance records. Series expansion would then
-		// restore the occurrence (issue #290). Only a missing master (ErrNoRows)
-		// may skip those records.
-		if err != nil && !errors.Is(err, sql.ErrNoRows) {
-			return fmt.Errorf("get master: %w", err)
-		}
-		if err == nil {
-			existing := timeutil.ParseTimeList(storage.NullableToString(master.Exdates))
-			recIDTime, parseErr := timeutil.ParseRecurrenceID(td.RecurrenceID)
-			if parseErr != nil {
-				// A malformed recurrence_id cannot be excluded from the
-				// master. A soft-delete of the override would then restore the
-				// occurrence via series expansion. Return an error. The restore
-				// path treats the same parse failure as fatal.
-				return fmt.Errorf("parse recurrence_id %q: %w", td.RecurrenceID, parseErr)
-			}
-			existing = append(existing, recIDTime)
-			if err := qtx.UpdateTodoExdates(ctx, storage.UpdateTodoExdatesParams{
-				Exdates: storage.StringToNullable(timeutil.SerializeTimeList(existing)),
-				ID:      master.ID,
-			}); err != nil {
-				return fmt.Errorf("update exdates: %w", err)
-			}
-			// Record provenance so restore knows this EXDATE was
-			// delete-added (and may be stripped) rather than imported.
-			if err := qtx.RecordTodoExdateDelete(ctx, storage.RecordTodoExdateDeleteParams{
-				CalendarID:   master.CalendarID,
-				Uid:          td.UID,
-				RecurrenceID: td.RecurrenceID,
-			}); err != nil {
-				return fmt.Errorf("record exdate delete: %w", err)
-			}
-		}
-
-		if err := qtx.SoftDeleteTodo(ctx, id); err != nil {
-			return fmt.Errorf("soft-delete todo: %w", err)
-		}
-		// Mark the master dirty — its EXDATE was modified — inside the same
-		// transaction. A failed mark then rolls the EXDATE change back. The
-		// path does not commit a change that is never pushed (issue #107).
-		if err := storage.MarkResourceDirty(ctx, tx, td.CalendarID, td.UID, "todo"); err != nil {
-			return fmt.Errorf("mark resource dirty: %w", err)
-		}
-		return tx.Commit()
-	}
-
-	// Unreachable: RecurrenceID is either "" (handled above) or non-empty.
-	if err := s.q.SoftDeleteTodo(ctx, id); err != nil {
-		return err
-	}
-	_ = storage.MarkResourceDirty(ctx, s.db, td.CalendarID, td.UID, "todo")
-	return nil
-}
-
-// DeleteSeries soft-deletes a recurring master todo and every override
-// with its UID. A tombstone is created when the master is synced so
-// the next push sends DELETE to the server. The local rows stay in
-// place until purge so the user can restore them.
-func (s *Service) DeleteSeries(ctx context.Context, uid string) error {
-	// Resolve every calendar the UID spans before a transaction opens. A
-	// read-only or VEVENT-only collection then rejects the whole series
-	// delete up front. Distinct calendar IDs across all rows (not just the
-	// master) cover orphaned overrides and series-tail rows even after the
-	// master has been purged on its own.
-	if err := s.ensureSeriesWritable(ctx, uid); err != nil {
-		return err
-	}
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-
-	// Tombstone, dirty-mark, and soft-delete commit together. A failed
-	// sync-track write cannot leave a tombstone for a still-live series
-	// whose next sync would DELETE it from the server (issue #107). No
-	// master means there is nothing to track.
-	master, mErr := qtx.GetTodoByUID(ctx, uid)
-	// Only ErrNoRows means "no master to track". A genuine lookup error must
-	// abort. Otherwise the series is soft-deleted locally without a tombstone
-	// or dirty mark. It would then resurface on the next pull (issue #290).
-	if mErr != nil && !errors.Is(mErr, sql.ErrNoRows) {
-		return fmt.Errorf("get master: %w", mErr)
-	}
-	haveMaster := mErr == nil
-	if haveMaster {
-		if _, err := storage.CreateTombstoneIfSynced(ctx, tx, master.CalendarID, uid); err != nil {
-			return fmt.Errorf("create tombstone: %w", err)
-		}
-	}
-
-	if err := qtx.SoftDeleteTodosByUID(ctx, uid); err != nil {
-		return fmt.Errorf("soft-delete series: %w", err)
-	}
-	if haveMaster {
-		if err := storage.MarkResourceDirty(ctx, tx, master.CalendarID, uid, "todo"); err != nil {
-			return fmt.Errorf("mark resource dirty: %w", err)
-		}
-	}
-	return tx.Commit()
-}
-
 // ListOverridesByUID returns all override instances for a given UID.
 func (s *Service) ListOverridesByUID(ctx context.Context, uid string) ([]Todo, error) {
 	rows, err := s.q.ListTodoOverridesByUID(ctx, uid)
@@ -793,785 +625,6 @@ func (s *Service) ListOverridesByUID(ctx context.Context, uid string) ([]Todo, e
 		return nil, err
 	}
 	return fromStorageSlice(rows), nil
-}
-
-// Alarm CRUD
-
-func (s *Service) ListAlarms(ctx context.Context, todoID int64) ([]model.Alarm, error) {
-	return s.listAlarms(ctx, todoID, true)
-}
-
-// ListAlarmsLean returns a todo's alarms with attendees (needed to fire
-// EMAIL alarms) but without X-properties, which are round-trip-only and
-// never read at fire time.
-func (s *Service) ListAlarmsLean(ctx context.Context, todoID int64) ([]model.Alarm, error) {
-	return s.listAlarms(ctx, todoID, false)
-}
-
-// ListFireableAlarmsByTodoIDs fetches the fireable alarms for many todo
-// IDs in one query. It returns a map of the todo ID to its alarms. The
-// alarm check loop reads every todo on each tick, so a call per todo
-// costs one query per todo (issue #586). The query excludes a preserved
-// sync-only action.
-func (s *Service) ListFireableAlarmsByTodoIDs(ctx context.Context, todoIDs []int64) (map[int64][]model.Alarm, error) {
-	if len(todoIDs) == 0 {
-		return nil, nil
-	}
-	rows, err := s.q.ListFireableTodoAlarmsByTodoIDs(ctx, todoIDs)
-	if err != nil {
-		return nil, err
-	}
-	alarms, err := s.buildAlarms(ctx, rows, false)
-	if err != nil {
-		return nil, err
-	}
-	if len(alarms) == 0 {
-		return nil, nil
-	}
-	// fromStorageTodoAlarm maps the todo ID onto the EventID field of the
-	// shared alarm model, like the event service does with its own ID.
-	alarmMap := make(map[int64][]model.Alarm, len(todoIDs))
-	for _, a := range alarms {
-		alarmMap[a.EventID] = append(alarmMap[a.EventID], a)
-	}
-	return alarmMap, nil
-}
-
-func (s *Service) listAlarms(ctx context.Context, todoID int64, withXProps bool) ([]model.Alarm, error) {
-	rows, err := s.q.ListTodoAlarmsByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	return s.buildAlarms(ctx, rows, withXProps)
-}
-
-// buildAlarms turns todo_alarms rows into the shared alarm model. It
-// attaches the attendees, and the X-properties when the caller asks for
-// them. The per-todo read and the batch read share it.
-func (s *Service) buildAlarms(ctx context.Context, rows []storage.TodoAlarm, withXProps bool) ([]model.Alarm, error) {
-	if len(rows) == 0 {
-		return nil, nil
-	}
-	alarmIDs := make([]int64, len(rows))
-	for i, r := range rows {
-		alarmIDs[i] = r.ID
-	}
-	// Load failures propagate. Attendees feed content match in ReplaceAlarms.
-	// X-properties feed export and sync pushes. A silent degraded alarm set
-	// would corrupt merges or rewrite the server copy.
-	attRows, err := s.q.ListTodoAlarmAttendeesByAlarmIDs(ctx, alarmIDs)
-	if err != nil {
-		return nil, fmt.Errorf("load alarm attendees: %w", err)
-	}
-	attMap := make(map[int64][]model.AlarmAttendee, len(rows))
-	for _, ar := range attRows {
-		attMap[ar.AlarmID] = append(attMap[ar.AlarmID], model.AlarmAttendee{
-			ID: ar.ID, Email: ar.Email, Name: storage.NullableToString(ar.Name),
-		})
-	}
-	alarms := make([]model.Alarm, len(rows))
-	for i, r := range rows {
-		alarms[i] = fromStorageTodoAlarm(r)
-		alarms[i].Attendees = attMap[r.ID]
-	}
-	if withXProps {
-		if err := storage.AttachAlarmXProperties(ctx, s.q, storage.OwnerTypeTodoAlarm, alarms); err != nil {
-			return nil, err
-		}
-	}
-	return alarms, nil
-}
-
-// fromStorageTodoAlarm converts an alarm row to the model value. It maps
-// a malformed stored action to model.UnsupportedAlarmAction, like the
-// event service does. See model.NormalizeAlarmAction (issue #607).
-func fromStorageTodoAlarm(r storage.TodoAlarm) model.Alarm {
-	return model.Alarm{
-		ID: r.ID, EventID: r.TodoID,
-		UID:           storage.NullableToString(r.Uid),
-		Action:        model.NormalizeAlarmAction(r.Action),
-		TriggerValue:  r.TriggerValue,
-		Description:   storage.NullableToString(r.Description),
-		Summary:       storage.NullableToString(r.Summary),
-		Repeat:        int(r.Repeat),
-		Duration:      storage.NullableToString(r.Duration),
-		Related:       r.Related,
-		Acknowledged:  storage.NullableToString(r.Acknowledged),
-		AttachURI:     storage.NullableToString(r.AttachUri),
-		AttachBinary:  r.AttachBinary,
-		AttachFmtType: storage.NullableToString(r.AttachFmttype),
-	}
-}
-
-// ReplaceFireableAlarms replaces the fireable alarms of a todo and
-// carries the stored sync-only rows forward, like the event method of the
-// same name (issue #579). A caller that must delete such a row calls
-// ReplaceAlarms instead.
-// The read of the stored rows and the write share one transaction, like
-// the event method of the same name. A read on its own connection could
-// return a row a concurrent pull has already deleted.
-func (s *Service) ReplaceFireableAlarms(ctx context.Context, todoID int64, alarms []model.Alarm) error {
-	if s.tx != nil {
-		return s.replaceFireableAlarms(ctx, todoID, alarms)
-	}
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if err := s.WithTx(tx).replaceFireableAlarms(ctx, todoID, alarms); err != nil {
-		return err
-	}
-	return tx.Commit()
-}
-
-// replaceFireableAlarms carries the stored sync-only rows forward. The
-// caller supplies the transaction.
-func (s *Service) replaceFireableAlarms(ctx context.Context, todoID int64, alarms []model.Alarm) error {
-	stored, err := s.ListAlarms(ctx, todoID)
-	if err != nil {
-		return fmt.Errorf("list stored alarms: %w", err)
-	}
-	return s.ReplaceAlarms(ctx, todoID, model.KeepSyncOnlyAlarms(stored, alarms))
-}
-
-// ClearSyncOnlyAlarms deletes every stored alarm the engine cannot fire and
-// keeps the rest, like the event method of the same name (issue #593). The
-// read of the stored rows and the write share one transaction.
-func (s *Service) ClearSyncOnlyAlarms(ctx context.Context, todoID int64) error {
-	if s.tx != nil {
-		return s.clearSyncOnlyAlarms(ctx, todoID)
-	}
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	if err := s.WithTx(tx).clearSyncOnlyAlarms(ctx, todoID); err != nil {
-		return err
-	}
-	return tx.Commit()
-}
-
-// clearSyncOnlyAlarms keeps the fireable stored rows. The caller supplies
-// the transaction.
-func (s *Service) clearSyncOnlyAlarms(ctx context.Context, todoID int64) error {
-	stored, err := s.ListAlarms(ctx, todoID)
-	if err != nil {
-		return fmt.Errorf("list stored alarms: %w", err)
-	}
-	return s.ReplaceAlarms(ctx, todoID, model.FireableAlarmsOnly(stored))
-}
-
-func (s *Service) ReplaceAlarms(ctx context.Context, todoID int64, alarms []model.Alarm) error {
-	td, err := s.Get(ctx, todoID)
-	if err != nil {
-		return err
-	}
-	if err := calendaraccess.EnsureWritable(ctx, s.q, td.CalendarID, todoComponent); err != nil {
-		return err
-	}
-	return s.ReplaceAlarmsForSync(ctx, todoID, alarms)
-}
-
-// ReplaceAlarmsForSync applies an alarm set without the remote access and
-// component policy. It is reserved for the CalDAV sync engine, which
-// mirrors a server-originated VTODO into the local cache whatever the
-// linked collection advertises. A user-originated edit must route through
-// ReplaceAlarms, so the policy holds.
-func (s *Service) ReplaceAlarmsForSync(ctx context.Context, todoID int64, alarms []model.Alarm) error {
-	// Prepare before the transaction opens. A standalone call then
-	// rejects a bad alarm without a write lock. A sync caller already
-	// holds its own transaction. See model.PrepareAlarmsForWrite.
-	alarms, err := model.PrepareAlarmsForWrite(alarms)
-	if err != nil {
-		return err
-	}
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-
-	// Merge. Do not wipe and recreate. A delete of a todo alarm row cascades
-	// away its todo_alarm_state. That would restore dismissed firings on
-	// every sync rewrite. Mirrors event.Service.ReplaceAlarms.
-	existing, err := loadExistingTodoAlarms(ctx, qtx, todoID)
-	if err != nil {
-		return err
-	}
-
-	matched := make([]bool, len(existing))
-	var unmatched []model.Alarm
-	for _, a := range alarms {
-		if j, found := matchTodoAlarm(existing, matched, a); found {
-			matched[j] = true
-			if err := syncMatchedTodoAlarm(ctx, qtx, a, existing[j]); err != nil {
-				return err
-			}
-		} else {
-			unmatched = append(unmatched, a)
-		}
-	}
-
-	// Second pass: content changed but the RFC 9074 UID is stable — the same
-	// alarm edited. Update in place so the row ID (and its state) survives.
-	for _, a := range unmatched {
-		if j, found := matchTodoAlarmByUID(existing, matched, a); found {
-			matched[j] = true
-			if err := updateTodoAlarmInPlace(ctx, qtx, todoID, a, existing[j]); err != nil {
-				return err
-			}
-		} else {
-			if err := createNewTodoAlarm(ctx, qtx, todoID, a); err != nil {
-				return err
-			}
-		}
-	}
-
-	for j, ex := range existing {
-		if !matched[j] {
-			if err := qtx.DeleteTodoAlarmByID(ctx, ex.ID); err != nil {
-				return fmt.Errorf("delete unmatched alarm: %w", err)
-			}
-		}
-	}
-
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
-}
-
-// loadExistingTodoAlarms loads a todo's alarms with attendees and
-// X-properties inside the transaction for merge match.
-func loadExistingTodoAlarms(ctx context.Context, qtx *storage.Queries, todoID int64) ([]model.Alarm, error) {
-	rows, err := qtx.ListTodoAlarmsByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, fmt.Errorf("list existing alarms: %w", err)
-	}
-	if len(rows) == 0 {
-		return nil, nil
-	}
-	alarmIDs := make([]int64, len(rows))
-	for i, r := range rows {
-		alarmIDs[i] = r.ID
-	}
-	attRows, err := qtx.ListTodoAlarmAttendeesByAlarmIDs(ctx, alarmIDs)
-	if err != nil {
-		return nil, fmt.Errorf("list alarm attendees: %w", err)
-	}
-	attMap := make(map[int64][]model.AlarmAttendee, len(rows))
-	for _, ar := range attRows {
-		attMap[ar.AlarmID] = append(attMap[ar.AlarmID], model.AlarmAttendee{
-			ID: ar.ID, Email: ar.Email, Name: storage.NullableToString(ar.Name),
-		})
-	}
-	alarms := make([]model.Alarm, len(rows))
-	for i, r := range rows {
-		alarms[i] = fromStorageTodoAlarm(r)
-		alarms[i].Attendees = attMap[r.ID]
-	}
-	if err := storage.AttachAlarmXProperties(ctx, qtx, storage.OwnerTypeTodoAlarm, alarms); err != nil {
-		return nil, err
-	}
-	return alarms, nil
-}
-
-// matchTodoAlarm tries to match an alarm that arrives with stored ones by
-// content. Rows whose non-empty RFC 9074 UIDs differ are never paired.
-// See event.matchAlarm for the rationale.
-func matchTodoAlarm(existing []model.Alarm, matched []bool, a model.Alarm) (int, bool) {
-	for j, ex := range existing {
-		if matched[j] {
-			continue
-		}
-		if a.UID != "" && ex.UID != "" && a.UID != ex.UID {
-			continue
-		}
-		if a.ContentEqual(ex) {
-			return j, true
-		}
-	}
-	return 0, false
-}
-
-// matchTodoAlarmByUID matches an alarm that arrives against unmatched stored
-// ones by RFC 9074 UID.
-func matchTodoAlarmByUID(existing []model.Alarm, matched []bool, a model.Alarm) (int, bool) {
-	if a.UID == "" {
-		return 0, false
-	}
-	for j, ex := range existing {
-		if matched[j] || ex.UID == "" {
-			continue
-		}
-		if ex.UID == a.UID {
-			return j, true
-		}
-	}
-	return 0, false
-}
-
-// syncMatchedTodoAlarm syncs a content-matched alarm's UID and ACKNOWLEDGED state.
-func syncMatchedTodoAlarm(ctx context.Context, qtx *storage.Queries, a model.Alarm, ex model.Alarm) error {
-	// Backfill the UID of a stored row that carries none. A preserved
-	// foreign alarm keeps its empty UID, so the backfill writes nothing
-	// for it (issue #586).
-	if uid := model.AlarmUIDForWrite(a); ex.UID == "" && uid != "" {
-		if err := qtx.UpdateTodoAlarmUID(ctx, storage.UpdateTodoAlarmUIDParams{
-			Uid: storage.StringToNullable(uid),
-			ID:  ex.ID,
-		}); err != nil {
-			return fmt.Errorf("backfill alarm uid: %w", err)
-		}
-	}
-	if a.Acknowledged != ex.Acknowledged && model.ValidateAcknowledged(a.Acknowledged) {
-		if err := qtx.UpdateTodoAlarmAcknowledged(ctx, storage.UpdateTodoAlarmAcknowledgedParams{
-			Acknowledged: storage.StringToNullable(a.Acknowledged),
-			ID:           ex.ID,
-		}); err != nil {
-			return fmt.Errorf("update alarm acknowledged: %w", err)
-		}
-	}
-	// X-properties are excluded from content match. Refresh them so a
-	// remote X-prop change still lands. nil means the caller has no X-prop
-	// knowledge. Keep the stored rows. Only a non-nil slice is authoritative.
-	if a.XProperties == nil || model.XPropsContentEqual(a.XProperties, ex.XProperties) {
-		return nil
-	}
-	return storage.ReplaceAlarmXProperties(ctx, qtx, storage.OwnerTypeTodoAlarm, ex.ID, a.XProperties)
-}
-
-// updateTodoAlarmInPlace rewrites a UID-matched alarm's content on its
-// stored row. The row ID stays so todo_alarm_state entries survive.
-func updateTodoAlarmInPlace(ctx context.Context, qtx *storage.Queries, todoID int64, a model.Alarm, ex model.Alarm) error {
-	// A rewrite to a sync-only action disables the alarm, but this code
-	// leaves todo_alarm_state alone. The pending and snooze queries
-	// filter on the current action, so the state of the alarm comes back
-	// when a later pull restores a fireable action (issue #579).
-	ack, err := model.PrepareAlarmUpdate(a, ex)
-	if err != nil {
-		return err
-	}
-	if err := qtx.UpdateTodoAlarmContentByID(ctx, storage.UpdateTodoAlarmContentByIDParams{
-		Action:        a.Action,
-		TriggerValue:  a.TriggerValue,
-		Description:   storage.StringToNullable(a.Description),
-		Summary:       storage.StringToNullable(a.Summary),
-		Repeat:        int64(a.Repeat),
-		Duration:      storage.StringToNullable(a.Duration),
-		Related:       a.Related,
-		Acknowledged:  storage.StringToNullable(ack),
-		AttachUri:     storage.StringToNullable(a.AttachURI),
-		AttachFmttype: storage.StringToNullable(a.AttachFmtType),
-		AttachBinary:  a.AttachBinary,
-		ID:            ex.ID,
-		TodoID:        todoID,
-	}); err != nil {
-		return fmt.Errorf("update alarm content: %w", err)
-	}
-	if err := qtx.DeleteTodoAlarmAttendeesByAlarmID(ctx, ex.ID); err != nil {
-		return fmt.Errorf("delete alarm attendees: %w", err)
-	}
-	for _, att := range a.Attendees {
-		_, err := qtx.CreateTodoAlarmAttendee(ctx, storage.CreateTodoAlarmAttendeeParams{
-			AlarmID: ex.ID,
-			Email:   att.Email,
-			Name:    storage.StringToNullable(att.Name),
-		})
-		if err != nil {
-			return fmt.Errorf("create alarm attendee: %w", err)
-		}
-	}
-	if a.XProperties == nil || model.XPropsContentEqual(a.XProperties, ex.XProperties) {
-		return nil
-	}
-	return storage.ReplaceAlarmXProperties(ctx, qtx, storage.OwnerTypeTodoAlarm, ex.ID, a.XProperties)
-}
-
-// isUniqueUIDViolation reports whether an insert failed on the global
-// alarm-UID unique index.
-func isUniqueUIDViolation(err error) bool {
-	if err == nil {
-		return false
-	}
-	msg := err.Error()
-	return strings.Contains(msg, "UNIQUE constraint failed") && strings.Contains(msg, ".uid")
-}
-
-// createNewTodoAlarm creates a new todo alarm and its attendees. On a global
-// UID collision (for example a server that duplicates a todo with its VALARM
-// UIDs), mint a fresh local UID. Do not fail the sync forever.
-func createNewTodoAlarm(ctx context.Context, qtx *storage.Queries, todoID int64, a model.Alarm) error {
-	if err := model.CheckStorableAlarmAction(a.Action); err != nil {
-		return fmt.Errorf("create alarm: %w", err)
-	}
-	uid := model.AlarmUIDForWrite(a)
-	params := storage.CreateTodoAlarmParams{
-		TodoID:        todoID,
-		Uid:           storage.StringToNullable(uid),
-		Action:        a.Action,
-		TriggerValue:  a.TriggerValue,
-		Description:   storage.StringToNullable(a.Description),
-		Summary:       storage.StringToNullable(a.Summary),
-		Repeat:        int64(a.Repeat),
-		Duration:      storage.StringToNullable(a.Duration),
-		Related:       a.Related,
-		Acknowledged:  storage.StringToNullable(a.Acknowledged),
-		AttachUri:     storage.StringToNullable(a.AttachURI),
-		AttachFmttype: storage.StringToNullable(a.AttachFmtType),
-		AttachBinary:  a.AttachBinary,
-	}
-	row, err := qtx.CreateTodoAlarm(ctx, params)
-	if isUniqueUIDViolation(err) {
-		// Retry through the shared rule, like the event service does.
-		// The retry must not stamp a minted UID on a preserved foreign
-		// alarm (issue #586).
-		retry := a
-		retry.UID = ""
-		params.Uid = storage.StringToNullable(model.AlarmUIDForWrite(retry))
-		row, err = qtx.CreateTodoAlarm(ctx, params)
-	}
-	if err != nil {
-		return fmt.Errorf("create alarm: %w", err)
-	}
-	for _, att := range a.Attendees {
-		_, err := qtx.CreateTodoAlarmAttendee(ctx, storage.CreateTodoAlarmAttendeeParams{
-			AlarmID: row.ID,
-			Email:   att.Email,
-			Name:    storage.StringToNullable(att.Name),
-		})
-		if err != nil {
-			return fmt.Errorf("create alarm attendee: %w", err)
-		}
-	}
-	if len(a.XProperties) == 0 {
-		return nil
-	}
-	return storage.ReplaceAlarmXProperties(ctx, qtx, storage.OwnerTypeTodoAlarm, row.ID, a.XProperties)
-}
-
-// Attendee CRUD
-
-func (s *Service) ListAttendees(ctx context.Context, todoID int64) ([]model.Attendee, error) {
-	rows, err := s.q.ListTodoAttendeesByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	attendees := make([]model.Attendee, len(rows))
-	for i, r := range rows {
-		attendees[i] = model.Attendee{
-			ID: r.ID, EventID: r.TodoID,
-			Email: r.Email, Name: storage.NullableToString(r.Name),
-			RSVPStatus: r.RsvpStatus, Role: r.Role,
-			Organizer:     r.Organizer == 1,
-			CUType:        storage.NullableToString(r.Cutype),
-			RSVPRequested: strings.EqualFold(storage.NullableToString(r.Rsvp), "TRUE"),
-			SentBy:        storage.NullableToString(r.SentBy),
-			DelegatedTo:   storage.NullableToString(r.DelegatedTo),
-			DelegatedFrom: storage.NullableToString(r.DelegatedFrom),
-			Member:        storage.NullableToString(r.Member),
-			Dir:           storage.NullableToString(r.Dir),
-			Language:      storage.NullableToString(r.Language),
-		}
-	}
-	return attendees, nil
-}
-
-func (s *Service) ReplaceAttendees(ctx context.Context, todoID int64, attendees []model.Attendee) error {
-	// Prepare before the transaction opens. A standalone call then
-	// rejects a bad attendee without a write lock. See
-	// model.PrepareAttendeesForWrite.
-	attendees, err := model.PrepareAttendeesForWrite(model.TaskAttendee, attendees)
-	if err != nil {
-		return err
-	}
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-
-	if err := qtx.DeleteTodoAttendeesByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete attendees: %w", err)
-	}
-	for _, a := range attendees {
-		org := storage.BoolToInt(a.Organizer)
-		rsvp := ""
-		if a.RSVPRequested {
-			rsvp = "TRUE"
-		}
-		_, err := qtx.CreateTodoAttendee(ctx, storage.CreateTodoAttendeeParams{
-			TodoID:        todoID,
-			Email:         a.Email,
-			Name:          storage.StringToNullable(a.Name),
-			RsvpStatus:    a.RSVPStatus,
-			Role:          a.Role,
-			Organizer:     org,
-			Cutype:        storage.StringToNullable(a.CUType),
-			Rsvp:          storage.StringToNullable(rsvp),
-			SentBy:        storage.StringToNullable(a.SentBy),
-			DelegatedTo:   storage.StringToNullable(a.DelegatedTo),
-			DelegatedFrom: storage.StringToNullable(a.DelegatedFrom),
-			Member:        storage.StringToNullable(a.Member),
-			Dir:           storage.StringToNullable(a.Dir),
-			Language:      storage.StringToNullable(a.Language),
-		})
-		if err != nil {
-			return fmt.Errorf("create attendee: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
-}
-
-// Category CRUD
-
-func (s *Service) ListCategories(ctx context.Context, todoID int64) ([]string, error) {
-	rows, err := s.q.ListCategoriesByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Category
-	}
-	return out, nil
-}
-
-func (s *Service) ListAllCategories(ctx context.Context) ([]string, error) {
-	return s.q.ListAllTodoCategories(ctx)
-}
-
-func (s *Service) ReplaceCategories(ctx context.Context, todoID int64, categories []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-
-	if err := replaceCategoriesTx(ctx, qtx, todoID, categories); err != nil {
-		return err
-	}
-	if err := commit(); err != nil {
-		return fmt.Errorf("commit replace categories: %w", err)
-	}
-	return nil
-}
-
-// replaceCategoriesTx wipes and rewrites a todo's categories using a tx-bound
-// Queries. It opens no transaction and does not commit, so callers can compose
-// it with the parent todo write inside one transaction (issue #222).
-func replaceCategoriesTx(ctx context.Context, qtx *storage.Queries, todoID int64, categories []string) error {
-	if err := qtx.DeleteCategoriesByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete categories: %w", err)
-	}
-	for _, c := range categories {
-		if _, err := qtx.CreateTodoCategory(ctx, storage.CreateTodoCategoryParams{
-			TodoID:   todoID,
-			Category: c,
-		}); err != nil {
-			return fmt.Errorf("create category: %w", err)
-		}
-	}
-	return nil
-}
-
-// Attachment CRUD
-
-func (s *Service) ListAttachments(ctx context.Context, todoID int64) ([]model.Attachment, error) {
-	rows, err := s.q.ListTodoAttachmentsByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]model.Attachment, len(rows))
-	for i, r := range rows {
-		out[i] = model.Attachment{ID: r.ID, URI: storage.NullableToString(r.Uri), FmtType: storage.NullableToString(r.Fmttype), Data: r.Data, Filename: storage.NullableToString(r.Filename)}
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceAttachments(ctx context.Context, todoID int64, attachments []model.Attachment) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteTodoAttachmentsByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete attachments: %w", err)
-	}
-	for _, a := range attachments {
-		_, err := qtx.CreateTodoAttachment(ctx, storage.CreateTodoAttachmentParams{
-			TodoID: todoID, Uri: storage.StringToNullable(a.URI), Fmttype: storage.StringToNullable(a.FmtType), Data: a.Data, Filename: storage.StringToNullable(a.Filename),
-		})
-		if err != nil {
-			return fmt.Errorf("create attachment: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
-}
-
-// Comment CRUD
-
-func (s *Service) ListComments(ctx context.Context, todoID int64) ([]string, error) {
-	rows, err := s.q.ListTodoCommentsByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Text
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceComments(ctx context.Context, todoID int64, comments []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteTodoCommentsByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete comments: %w", err)
-	}
-	for _, c := range comments {
-		_, err := qtx.CreateTodoComment(ctx, storage.CreateTodoCommentParams{
-			TodoID: todoID, Text: c,
-		})
-		if err != nil {
-			return fmt.Errorf("create comment: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
-}
-
-// Contact CRUD
-
-func (s *Service) ListContacts(ctx context.Context, todoID int64) ([]string, error) {
-	rows, err := s.q.ListTodoContactsByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Text
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceContacts(ctx context.Context, todoID int64, contacts []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteTodoContactsByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete contacts: %w", err)
-	}
-	for _, c := range contacts {
-		_, err := qtx.CreateTodoContact(ctx, storage.CreateTodoContactParams{
-			TodoID: todoID, Text: c,
-		})
-		if err != nil {
-			return fmt.Errorf("create contact: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
-}
-
-// Resource CRUD
-
-func (s *Service) ListResources(ctx context.Context, todoID int64) ([]string, error) {
-	rows, err := s.q.ListTodoResourcesByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]string, len(rows))
-	for i, r := range rows {
-		out[i] = r.Text
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceResources(ctx context.Context, todoID int64, resources []string) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteTodoResourcesByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete resources: %w", err)
-	}
-	for _, r := range resources {
-		_, err := qtx.CreateTodoResource(ctx, storage.CreateTodoResourceParams{
-			TodoID: todoID, Text: r,
-		})
-		if err != nil {
-			return fmt.Errorf("create resource: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
-}
-
-// Relation CRUD
-
-func (s *Service) ListRelations(ctx context.Context, todoID int64) ([]model.Relation, error) {
-	rows, err := s.q.ListTodoRelationsByTodoID(ctx, todoID)
-	if err != nil {
-		return nil, err
-	}
-	out := make([]model.Relation, len(rows))
-	for i, r := range rows {
-		out[i] = model.Relation{ID: r.ID, RelType: r.RelType, RelUID: r.RelUid}
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceRelations(ctx context.Context, todoID int64, relations []model.Relation) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteTodoRelationsByTodoID(ctx, todoID); err != nil {
-		return fmt.Errorf("delete relations: %w", err)
-	}
-	for _, r := range relations {
-		_, err := qtx.CreateTodoRelation(ctx, storage.CreateTodoRelationParams{
-			TodoID: todoID, RelType: r.RelType, RelUid: r.RelUID,
-		})
-		if err != nil {
-			return fmt.Errorf("create relation: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
 }
 
 // Converters
@@ -1645,51 +698,6 @@ func (s *Service) populateCategories(ctx context.Context, todos []Todo) {
 			todos[i].Categories = timeutil.JoinCategoryList(cats)
 		}
 	}
-}
-
-// X-Property CRUD
-
-func (s *Service) ListXProperties(ctx context.Context, todoID int64) ([]model.XProperty, error) {
-	rows, err := s.q.ListXPropertiesByOwner(ctx, storage.ListXPropertiesByOwnerParams{
-		OwnerType: "todo", OwnerID: todoID,
-	})
-	if err != nil {
-		return nil, err
-	}
-	out := make([]model.XProperty, len(rows))
-	for i, r := range rows {
-		out[i] = model.XProperty{
-			ID: r.ID, OwnerType: r.OwnerType, OwnerID: r.OwnerID,
-			Name: r.Name, Value: r.Value, Params: r.Params,
-		}
-	}
-	return out, nil
-}
-
-func (s *Service) ReplaceXProperties(ctx context.Context, todoID int64, xprops []model.XProperty) error {
-	qtx, commit, rollback, err := s.txscope(ctx)
-	if err != nil {
-		return err
-	}
-	defer rollback()
-	if err := qtx.DeleteXPropertiesByOwner(ctx, storage.DeleteXPropertiesByOwnerParams{
-		OwnerType: "todo", OwnerID: todoID,
-	}); err != nil {
-		return fmt.Errorf("delete x-properties: %w", err)
-	}
-	for _, xp := range xprops {
-		if err := qtx.InsertXProperty(ctx, storage.InsertXPropertyParams{
-			OwnerType: "todo", OwnerID: todoID,
-			Name: xp.Name, Value: xp.Value, Params: xp.Params,
-		}); err != nil {
-			return fmt.Errorf("insert x-property: %w", err)
-		}
-	}
-	if err := commit(); err != nil {
-		return err
-	}
-	s.markDirtyByID(ctx, todoID)
-	return nil
 }
 
 func fromStorageSlice(rows []storage.Todo) []Todo {
